hash asset filenames for cache busting
Configure gatsby-plugin-manifest to include the hash of the icons in the
filenames. By default, it uses query strings for cache busting, but this
is inconsistent as the rest of gatsby appends the hash in the assets
filenames.

Another reason is that [using query strings is not recommended due to the
way some older proxies work][1] (the article is from 2008 so it might be
outdated, but then again so is the infrastructure in several parts of the
world).

[1]: http://www.stevesouders.com/blog/2008/08/23/revving-filenames-dont-use-querystring/